+++ CrossSolar/Models/OneDayElectricityModel.cs
@@ -9,13 +9,13 @@ namespace CrossSolar.Domain
     public class OneDayElectricityModel
     {
 
-        public double Sum { get; set; }
+        public decimal Sum { get; set; }
 
-        public double Average { get; set; }
+        public decimal Average { get; set; }
 
-        public double Maximum { get; set; }
+        public decimal Maximum { get; set; }
 
-        public double Minimum { get; set; }
+        public decimal Minimum { get; set; }
 
         public DateTime DateTime { get; set; }
     }
